code review

+++ lib/tasks/pull_failures_handler.rake
@@ -12,29 +12,37 @@ namespace :sync do
        # fetch records from db
        failed_files = FailedFiles.all
        failed_files.each do |record|
-          if download_file?(record.file_url, record.output_file_name, record.file_type)
-            object = record.object_type.constantize.find(record.object_id)
-            
-            # delete old logo
-            record_file_name = record.output_file_name
-            file_type = record_file_name[record_file_name.rindex(".")+1, record_file_name.length]
-            file_name = record.object_type.downcase.pluralize
-            delete_file(object, file_name, file_type)            
-            
-            # update object parameters
-            failed_file_parameters =  FailedFilesParameters.where(:failed_files_id => record.id)
-            failed_file_parameters.each do |file_paramters|
-              object[file_paramters.parameter] = file_paramters.value
-            end
-            object.save
-            
-            # upload file
-            upload_file(object)  
-                 
+          object = record.object_type.constantize.find(record.object_id)
+          failed_file_parameters =  FailedFilesParameters.where(:failed_files_id => record.id)
+          if object.updated_at > record.created_at
             #delete this record           
             record.destroy
             failed_file_parameters.each do |file_paramters|
              file_paramters.destroy
+            end
+          else
+            if download_file?(record.file_url, record.output_file_name, record.file_type)
+              
+              # delete old logo
+              record_file_name = record.output_file_name
+              file_type = record_file_name[record_file_name.rindex(".")+1, record_file_name.length]
+              file_name = record.object_type.downcase.pluralize
+              delete_file(object, file_name, file_type)            
+              
+              # update object parameters
+              failed_file_parameters.each do |file_paramters|
+                object[file_paramters.parameter] = file_paramters.value
+              end
+              object.save
+              
+              # upload file
+              upload_file(object)  
+                   
+              #delete this record           
+              record.destroy
+              failed_file_parameters.each do |file_paramters|
+               file_paramters.destroy
+              end
             end
           end
         end
